requirements.txt
Django
djangorestframework
pymongo
djangorestframework-simplejwt
djongo
python-decouple
Pillow
requests
social-auth-app-django
boto3
gunicorn
django-redis
django-cors-headers
redis
